requirements.txt
jupyter
ipykernel
notebook
colorama
ipdb
matplotlib
seaborn
numpy
pandas
scipy
scikit-learn
# API
fastapi
uvicorn
pydantic
python-dotenv

#SQL
pyodbc
